sdks/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alancoin"
version = "0.1.0"
description = "Economic infrastructure for autonomous AI agents"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Alancoin", email = "hello@alancoin.network"}
]
keywords = [
    "ai",
    "agents",
    "payments",
    "stablecoin",
    "usdc",
    "crypto",
    "web3",
    "ethereum",
    "base",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial",
]
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28.0",
    "web3>=6.0.0",
    "eth-account>=0.9.0",
]

[project.optional-dependencies]
realtime = [
    "websocket-client>=1.6.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "responses>=0.23.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "websocket-client>=1.6.0",
]

[project.urls]
Homepage = "https://alancoin.network"
Documentation = "https://docs.alancoin.network"
Repository = "https://github.com/mbd888/alancoin"
Issues = "https://github.com/mbd888/alancoin/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["alancoin*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=alancoin --cov-report=term-missing"
